+++ src/Products/MeetingLalouviere/tests/testCustomViews.py
@@ -2,6 +2,7 @@
 from datetime import datetime
 
 from Products.MeetingCommunes.tests.testCustomViews import testCustomViews as mctcv
+from Products.MeetingLalouviere.browser.overrides import MLLItemDocumentGenerationHelperView
 from Products.MeetingLalouviere.browser.overrides import MLLMeetingDocumentGenerationHelperView
 from Products.MeetingLalouviere.tests.MeetingLalouviereTestCase import (
     MeetingLalouviereTestCase,
@@ -19,6 +20,14 @@ class testCustomViews(mctcv, MeetingLalouviereTestCase):
         view = meeting.restrictedTraverse("@@document-generation")
         helper = view.get_generation_context_helper()
         self.assertTrue(isinstance(helper, MLLMeetingDocumentGenerationHelperView))
+
+    def test_MLLItemDocumentGenerationHelperView(self):
+        """Test if the browser layer is correctly applied"""
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem')
+        view = item.restrictedTraverse("@@document-generation")
+        helper = view.get_generation_context_helper()
+        self.assertTrue(isinstance(helper, MLLItemDocumentGenerationHelperView))
 
 
 def test_suite():